// src/types/common.ts
export type SelectOption = {
  value: string;
  label: string;
  disabled?: boolean;
  group?: string;
};